implementaion of chulcheck program
I implement chulcheck program using java-swing
So, there are some functions that are student registration,  subject registration, registering for courses and checking attendance by subject.
And all data is stored in a text file. Consequently, when you run the program, the data remains.
It is somewhat lacking, but I have implemented interestingly the program with ui interestingly.